Move class
Moved the Circumcircle class to the Triangle module since that's where it is going to be used.
